Enable microphone level meter
Mostly naively using some of the code from the soundmeter pip package

Right now I am assuming 2000 is the max rms level and hence percentages are based on that. However, I believe it's possible to get rms of greater than 200 and hence the graphical meter looks wonky. For now, leaving it as-is until I can figure out a better strategy for visualizing the rms value.